+++ CHANGELOG.md
@@ -334,3 +334,4 @@
 [@DmitryTsepelev]: https://github.com/DmitryTsepelev
 [@korolvs]: https://github.com/korolvs
 [@nicolas-brousse]: https://github.com/nicolas-brousse
+[@somenugget]: https://github.com/somenugget